Fix bug and make the implementation uniform with windows impl
